pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scrnaseq-hodge-pipeline"
version = "1.0.0"
description = "scRNAseq analysis pipeline using discrete Hodge decomposition to identify upstream cell types and genes"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "anndata>=0.10",
    "scanpy>=1.9",
    "numpy>=1.24",
    "scipy>=1.10",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "statsmodels>=0.14",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
gpu = ["torch>=2.0"]
dev = ["pytest>=7.0", "ruff>=0.1"]

[project.scripts]
scrnaseq-hodge = "run_pipeline:main"

[tool.setuptools.packages.find]
include = ["scripts*"]

[tool.ruff]
line-length = 100
target-version = "py310"
